chore: coveralls integration
